feat(yaml): add support for explicit Failsafe schema
The only thing missing is local tag resolution at this point.

Refs #115
