+++ apidom/packages/apidom-ast/src/yaml/schemas/Tag.ts
@@ -1,0 +1,16 @@
+import stampit from 'stampit';
+
+import ScalarTag from './ScalarTag';
+
+const Tag = stampit(ScalarTag, {
+  props: {
+    tag: '',
+  },
+  methods: {
+    resolve(value) {
+      return value;
+    },
+  },
+});
+
+export default Tag;
