+++ apidom/packages/apidom-ast/src/yaml/schemas/failsafe/GenericSequence.ts
@@ -1,0 +1,25 @@
+import stampit from 'stampit';
+
+import Tag from '../Tag';
+import { YamlNodeKind } from '../../nodes/YamlTag';
+
+const GenericSequence = stampit(Tag, {
+  statics: {
+    uri: 'tag:yaml.org,2002:seq',
+  },
+  init(args, { stamp }) {
+    this.tag = stamp.uri;
+  },
+  methods: {
+    test(node) {
+      // @ts-ignore
+      return node.tag.kind === YamlNodeKind.Sequence;
+    },
+
+    resolve(node) {
+      return node;
+    },
+  },
+});
+
+export default GenericSequence;
