+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/folded-clip.yaml
