+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/literal-indentation-indicator-expected.yaml
