+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/literal-keep-expected.yaml
