+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/literal-strip.yaml
