+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/plain-expected.yaml
