+++ apidom/packages/apidom-ast/test/yaml/schemas/failsafe/canonical-format/fixtures/single-quoted-expected.yaml
